Move PSNR metrics to src/evaluations.py
- src/evaluations.py: compute_psnr (full + masked) and compute_psnr_stratified
  (per-damage-type breakdown by mask channel)
- src/inference.py: remove compute_psnr, drop unused Optional import
- src/train.py: import compute_psnr from evaluations instead of inference
- README: update project structure
